refactor: remove the `move-vm-types` dependency from `iota-types`
