+++ crates/iota-types/Cargo.toml
@@ -73,7 +73,6 @@ move-disassembler.workspace = true
 move-ir-types.workspace = true
 move-vm-profiler.workspace = true
 move-vm-test-utils.workspace = true
-move-vm-types.workspace = true
 shared-crypto.workspace = true
 typed-store-error.workspace = true
